requirements.txt
datetime==4.3
fastapi==0.70.0
uvicorn==0.7.2
python-jose==3.3.0
pydantic==1.8.2
passlib==1.7.4
typing==3.10.0.0